+++ changelog.md
@@ -23,6 +23,7 @@
   anymore. The property `number_decimal_places` can be now set to `0` to indicate integers
   instead.
 * Fixed error when the select row modal is closed immediately after opening.
+* Hide "Export view" button if there is no valid exporter available
 
 ## Released (2022-01-13 1.8.2)